discord-bot/requirements.txt
inflect~=5.3.0
discord.py~=1.7.3